Feat: Smooth filtering thorught make page.js server components #34
